Fix after CR #1
