--- upgrade/assets/upgrade-job-rbac.yaml
@@ -1,49 +1,0 @@
-kind: ServiceAccount
-apiVersion: v1
-metadata:
-  name: upgrade-test-job-account
-  namespace: catalog
-  labels:
-    upgrade-test-job: "true"
-
----
-kind: ClusterRole
-apiVersion: rbac.authorization.k8s.io/v1beta1
-metadata:
-  name: upgrade-test-job-account
-  labels:
-    upgrade-test-job: "true"
-rules:
-  - apiGroups: [""]
-    resources: ["namespaces"]
-    verbs:     ["create"]
-  - apiGroups: ["apps"]
-    resources: ["deployments"]
-    verbs:     ["get"]
-  - apiGroups: ["servicecatalog.k8s.io"]
-    resources:
-      - "clusterserviceclasses"
-      - "clusterserviceplans"
-      - "clusterservicebrokers"
-      - "serviceinstances"
-      - "servicebindings"
-      - "serviceclasses"
-      - "serviceplans"
-      - "servicebrokers"
-    verbs: ["get", "list","create", "update", "delete"]
-
----
-kind: ClusterRoleBinding
-apiVersion: rbac.authorization.k8s.io/v1beta1
-metadata:
-  name: upgrade-test-job-account
-  labels:
-    upgrade-test-job: "true"
-subjects:
-  - kind: ServiceAccount
-    name: upgrade-test-job-account
-    namespace: catalog
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: upgrade-test-job-account
