--- upgrade/examiner/internal/signal.go
@@ -1,24 +1,0 @@
-package internal
-
-import (
-	"os"
-	"os/signal"
-	"syscall"
-)
-
-// SetupChannel registered for SIGTERM and Interrupt. A stop channel is returned
-// which is closed on one of these signals. If a second signal is caught, the program
-// is terminated with exit code 1.
-func SetupChannel() (stopCh <-chan struct{}) {
-	stop := make(chan struct{})
-	c := make(chan os.Signal, 2)
-	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
-	go func() {
-		<-c
-		close(stop)
-		<-c
-		os.Exit(1) // second signal. Exit directly.
-	}()
-
-	return stop
-}
